+++ src/internal/dots.c
@@ -1,9 +1,11 @@
 #include <rlang.h>
 #include "dots.h"
 #include "expr-interp.h"
+#include "internal.h"
 #include "utils.h"
 
 sexp* rlang_ns_get(const char* name);
+
 
 struct dots_capture_info {
   enum dots_capture_type type;
@@ -86,7 +88,6 @@ static sexp* def_unquote_name(sexp* expr, sexp* env) {
 
 
 static sexp* rlang_spliced_flag = NULL;
-
 static inline bool is_spliced_dots(sexp* x) {
   return r_get_attribute(x, rlang_spliced_flag) != r_null;
 }
@@ -94,8 +95,23 @@ static inline void mark_spliced_dots(sexp* x) {
   r_poke_attribute(x, rlang_spliced_flag, rlang_spliced_flag);
 }
 
-static sexp* dots_big_bang_coerce(sexp* expr) {
-  switch (r_typeof(expr)) {
+void signal_retired_splice() {
+  const char* msg =
+    "Unquoting language objects with `!!!` is soft-deprecated as of rlang 0.3.0.\n"
+    "Please use `!!` instead.\n"
+    "\n"
+    "  # Bad:\n"
+    "  dplyr::select(data, !!!enquo(x))\n"
+    "\n"
+    "  # Good:\n"
+    "  dplyr::select(data, !!enquo(x))    # Unquote single quosure\n"
+    "  dplyr::select(data, !!!enquos(x))  # Splice list of quosures\n";
+    r_signal_soft_deprecated(msg, msg, "rlang", r_empty_env);
+}
+
+// Maintain parity with deep_big_bang_coerce() in expr-interp.c
+static sexp* dots_big_bang_coerce(sexp* x) {
+  switch (r_typeof(x)) {
   case r_type_null:
   case r_type_pairlist:
   case r_type_logical:
@@ -104,18 +120,44 @@ static sexp* dots_big_bang_coerce(sexp* expr) {
   case r_type_complex:
   case r_type_character:
   case r_type_raw:
-    return r_vec_coerce(expr, r_type_list);
+    if (r_is_object(x)) {
+      return r_eval_with_x(as_list_call, r_base_env, x);
+    } else {
+      return r_vec_coerce(x, r_type_list);
+    }
   case r_type_list:
-    return r_duplicate(expr, true);
+    if (r_is_object(x)) {
+      return r_eval_with_x(as_list_call, r_base_env, x);
+    } else {
+      return r_duplicate(x, true);
+    }
+  case r_type_s4:
+    return r_eval_with_x(as_list_s4_call, r_methods_ns_env, x);
   case r_type_call:
-    if (r_is_symbol(r_node_car(expr), "{")) {
-      return r_vec_coerce(r_node_cdr(expr), r_type_list);
+    if (r_is_symbol(r_node_car(x), "{")) {
+      return r_vec_coerce(r_node_cdr(x), r_type_list);
     }
     // else fallthrough
+  case r_type_symbol:
+    signal_retired_splice();
+    return r_new_list(x, NULL);
+
   default:
-    return r_new_list(expr, NULL);
+    r_abort(
+      "Can't splice an object of type `%s` because it is not a vector",
+      r_type_as_c_string(r_typeof(x))
+    );
   }
 }
+
+static sexp* dots_value_big_bang(sexp* x) {
+  x = KEEP(dots_big_bang_coerce(x));
+  r_push_class(x, "spliced");
+
+  FREE(1);
+  return x;
+}
+
 static sexp* dots_big_bang(struct dots_capture_info* capture_info,
                            sexp* expr, sexp* env, bool quosured) {
   sexp* value = KEEP(r_eval(expr, env));
@@ -135,25 +177,6 @@ static sexp* dots_big_bang(struct dots_capture_info* capture_info,
 
   FREE(2);
   return value;
-}
-
-static sexp* set_value_spliced(sexp* x) {
-  static sexp* spliced_str = NULL;
-  if (!spliced_str) {
-    spliced_str = r_chr("spliced");
-    r_mark_precious(spliced_str);
-    r_mark_shared(spliced_str);
-  }
-
-  int n_protect = 0;
-
-  if (r_typeof(x) != r_type_list) {
-    x = KEEP_N(r_vec_coerce(x, r_type_list), n_protect);
-  }
-  x = r_set_class(x, spliced_str);
-
-  FREE(n_protect);
-  return x;
 }
 
 static inline bool should_ignore(int ignore_empty, r_ssize i, r_ssize n) {
@@ -255,16 +278,6 @@ static sexp* dots_unquote(sexp* dots, struct dots_capture_info* capture_info) {
     case OP_EXPR_UQS:
       capture_info->needs_expansion = true;
       expr = dots_big_bang(capture_info, info.operand, env, false);
-
-      // Work around bug in dplyr 0.7.4
-      int n = r_length(expr);
-      for (int i = 0; i < n; ++i) {
-        sexp* elt = r_list_get(expr, i);
-        if (rlang_is_quosure(elt)) {
-          r_list_poke(expr, i, rlang_quo_get_expr(elt));
-        }
-      }
-
       break;
     case OP_QUO_NONE:
     case OP_QUO_UQ:
@@ -308,7 +321,7 @@ static sexp* dots_unquote(sexp* dots, struct dots_capture_info* capture_info) {
       if (expr == r_null) {
         expr = empty_spliced_list();
       } else {
-        expr = set_value_spliced(expr);
+        expr = dots_value_big_bang(expr);
         capture_info->count += 1;
       }
       FREE(1);
@@ -371,16 +384,16 @@ static int find_auto_names_width(sexp* named) {
   r_abort("`.named` must be a scalar logical or number");
 }
 
+static sexp* auto_name_call = NULL;
+
 static sexp* maybe_auto_name(sexp* x, sexp* named) {
   int names_width = find_auto_names_width(named);
   sexp* names = r_vec_names(x);
 
   if (names_width && (!names || r_chr_has(names, ""))) {
-    sexp* auto_fn = KEEP(rlang_ns_get("quos_auto_name"));
     sexp* width = KEEP(r_int(names_width));
-    sexp* auto_call = KEEP(r_build_call2(auto_fn, x, width));
-    x = r_eval(auto_call, r_empty_env);
-    FREE(3);
+    x = r_eval_with_xy(auto_name_call, r_base_env, x, width);
+    FREE(1);
   }
 
   return x;
@@ -629,4 +642,9 @@ sexp* rlang_dots_flat_list(sexp* frame_env,
 
   FREE(1);
   return dots;
+}
+
+void rlang_init_dots() {
+  auto_name_call = r_parse("rlang:::quos_auto_name(x, y)");
+  r_mark_precious(auto_name_call);
 }