+++ tests/testthat/test-dots.R
@@ -196,4 +196,3 @@ test_that("dots collectors never warn for <- when option is set", {
   expect_no_warning(myexprs(a <- 1))
   expect_no_warning(myquos(a <- 1))
 })
-